Explicitly mark narrowing conversion

+++ src/rpc/xmlrpc_tinyxml2.cc
@@ -388,7 +388,7 @@ void XmlRpc::cleanup() {}
 void XmlRpc::insert_command(const char*, const char*, const char*) {}
 void XmlRpc::set_dialect(int) {}
 
-int64_t XmlRpc::size_limit() { return m_sizeLimit; }
+int64_t XmlRpc::size_limit() { return static_cast<int64_t>(m_sizeLimit); }
 void    XmlRpc::set_size_limit(uint64_t limit) { m_sizeLimit = limit; }
 
 bool    XmlRpc::is_valid() const { return m_isValid; }
